+++ managers/__init__.py
@@ -4,10 +4,8 @@ Gestores auxiliares para GRK PowerSloth
 
 from .localization_manager import LocalizationManager
 from .theme_manager import ThemeManager
-from .notification_manager import NotificationManager
 
 __all__ = [
     'LocalizationManager',
     'ThemeManager',
-    'NotificationManager'
 ]